move player code to player.go
