+++ main.go
@@ -2,16 +2,11 @@ package main
 
 import (
 	"context"
-	"database/sql"
 	"errors"
-	"math/rand"
-	"net/http"
 	"log"
-	"time"
+	"net/http"
 	"strconv"
-	"encoding/json"
-	"os"
-	"io/ioutil"
+	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/gorilla/websocket"
@@ -26,35 +21,6 @@ import (
 
 const (
 	DbName = "sudoku"
-	PlayerIdCtxKey = "playerId"
-)
-
-type (
-	PlayerColor string
-	PlayerName string
-)
-
-var (
-	// Set of A100 from https://material.io/design/color/#tools-for-picking-colors
-	AssignablePlayerColors = []PlayerColor{
-		"#FF5252",
-		"#FF4081",
-		"#E040FB",
-		"#7C4DFF",
-		"#536DFE",
-		// "#448AFF", // main player color
-		"#40C4FF",
-		"#18FFFF",
-		"#64FFDA",
-		"#69F0AE",
-		"#B2FF59",
-		"#EEFF41",
-		"#FFFF00",
-		"#FFD740",
-		"#FFAB40",
-		"#FF6E40",
-	}
-	SuperheroNames []PlayerName
 )
 
 type Server struct {
@@ -67,15 +33,6 @@ type Game struct {
 	Data   string
 	Name   string
 	Solved bool
-}
-
-// Merged player + state for primary game.
-type Player struct {
-	Id   int64 `sql:",primary" graphql:",key"`
-	Name PlayerName
-	Color PlayerColor
-	X int64
-	Y int64
 }
 
 type Message struct {
@@ -121,46 +78,12 @@ func (s *Server) registerGameQueries(schema *schemabuilder.Schema) {
 	})
 }
 
-// func (s *Server) resetPlayers(schema *schemabuilder.Schema) {
-// 	var result *Player
-// 	err := s.db.QueryRow(ctx, &result, sqlgen.Filter{"id": args.Id}, nil)
-// 	if err == sql.ErrNoRows {
-// 		return nil, nil
-// 	}
-// }
-
 func (s *Server) registerPlayerQueries(schema *schemabuilder.Schema) {
 	object := schema.Query()
 
-	object.FieldFunc("currentPlayer", func(ctx context.Context) (*Player, error) {
-		id := PlayerId(ctx)
-		log.Println("currentPlayer", id)
-		if id == nil {
-			return nil, nil
-		}
-		var result *Player
-		err := s.db.QueryRow(ctx, &result, sqlgen.Filter{"id": id}, nil)
-		if err == sql.ErrNoRows {
-			return nil, nil
-		}
-		return result, err
-	})
-	object.FieldFunc("player", func(ctx context.Context, args struct{ Id int64 }) (*Player, error) {
-		var result *Player
-		err := s.db.QueryRow(ctx, &result, sqlgen.Filter{"id": args.Id}, nil)
-		if err == sql.ErrNoRows {
-			return nil, nil
-		}
-		return result, err
-	})
-
-	object.FieldFunc("players", func(ctx context.Context) ([]*Player, error) {
-		var result []*Player
-		if err := s.db.Query(ctx, &result, nil, nil); err != nil {
-			return nil, err
-		}
-		return result, nil
-	})
+	object.FieldFunc("currentPlayer", s.GetCurrentPlayer)
+	object.FieldFunc("player", s.GetPlayerById)
+	object.FieldFunc("players", s.GetPlayers)
 }
 
 func (s *Server) registerGameMutations(schema *schemabuilder.Schema) {
@@ -218,25 +141,8 @@ func (s *Server) registerGameMutations(schema *schemabuilder.Schema) {
 func (s *Server) registerPlayerMutations(schema *schemabuilder.Schema) {
 	object := schema.Mutation()
 
-	object.FieldFunc("createPlayer", func(ctx context.Context, args struct{ Name PlayerName }) error {
-		_, err := s.db.InsertRow(ctx, &Player{Name: args.Name})
-		return err
-	})
-
-	type updatePlayerSelectionArgs struct {
-		PlayerId int64
-		X int64
-		Y int64
-	}
-	object.FieldFunc("updatePlayerSelection", func(ctx context.Context, args updatePlayerSelectionArgs) error {
-		player := &Player{}
-		if err := s.db.QueryRow(ctx, &player, sqlgen.Filter{"id": args.PlayerId}, nil); err != nil {
-			return err
-		}
-		player.X = args.X
-		player.Y = args.Y
-		return s.db.UpdateRow(ctx, player)
-	})
+	object.FieldFunc("createPlayer", s.CreatePlayer)
+	object.FieldFunc("updatePlayerSelection", s.UpdatePlayerSelection)
 }
 
 func (s *Server) registerMessageQuery(schema *schemabuilder.Schema) {
@@ -264,19 +170,8 @@ func (s *Server) registerMessageMutation(schema *schemabuilder.Schema) {
 	})
 }
 
-func PlayerId(ctx context.Context) *int64 {
-	val := ctx.Value(PlayerIdCtxKey)
-	if val == nil {
-		return nil
-	}
-	return val.(*int64)
-}
 
 func int64Ptr(i int64) *int64 { return &i }
-
-func WithPlayerId(ctx context.Context, playerId int64) context.Context {
-	return context.WithValue(ctx, PlayerIdCtxKey, int64Ptr(playerId))
-}
 
 func (s *Server) SchemaBuilderSchema() *schemabuilder.Schema {
 	schema := schemabuilder.NewSchema()
@@ -295,12 +190,6 @@ func (s *Server) Schema() *graphql.Schema {
 	return s.SchemaBuilderSchema().MustBuild()
 }
 
-func NewPlayer() *Player {
-	return &Player{
-		Name: SuperheroNames[rand.Intn(len(SuperheroNames))],
-		Color: AssignablePlayerColors[rand.Intn(len(AssignablePlayerColors))],
-	}
-}
 
 type executionLogger struct {}
 func (s *executionLogger) StartExecution(ctx context.Context, tags map[string]string, initial bool) {}
@@ -371,18 +260,6 @@ func handlerWithPlayerTracking(schema *graphql.Schema, server *Server) http.Hand
 }
 
 func main() {
-	superheroNamesFile, err := os.Open("./superheroes.json")
-	if err != nil {
-		panic(err)
-	}
-
-	superheroNamesByteValue, _ := ioutil.ReadAll(superheroNamesFile)
-
-	err = json.Unmarshal([]byte(superheroNamesByteValue), &SuperheroNames)
-	if err != nil {
-		panic(err)
-	}
-	superheroNamesFile.Close()
 
 	sqlgenSchema := sqlgen.NewSchema()
 	sqlgenSchema.MustRegisterType("games", sqlgen.AutoIncrement, Game{})
